pyproject.toml
[tool.poetry]
name = "quanweb"
version = "0.1.2"
description = "Quan's personal web blog"
authors = ["Nguyễn Hồng Quân <ng.hong.quan@gmail.com>"]
homepage = "http://quan.hoabinh.vn"

[tool.poetry.dependencies]
python = "^3.6"
Flask-Admin = "^1.5"
Flask-WTF = "^0.14.2"
Flask-Login = "^0.4.1"
Flask-Script = "^2.0"
Flask-Bootstrap = "^3.3"
LogentriesLogger = "^0.2.1"
blinker = "^1.4"
htmllaundry = "^2.1"
mdx_linkify = "^1.2"
gunicorn = "^19.9"
awesome-slugify = "^1.6"
html3 = "^1.17"
Flask-SQLAlchemy = "^2.3"
flask-behind-proxy = "^0.1.0"
Babel = "^2.6"
mistletoe = "^0.7.2"
SQLAlchemy-Utils = "^0.35.0"
meinheld = "^1.0.1"
arrow = "^0.15.4"
alembic = "^1.3.1"
psycopg2-binary = "^2.8.4"
Pygments = "^2.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
